+++ app/src/main/java/com/github/pwittchen/reactivenetwork/app/MainActivity.java
@@ -20,10 +20,10 @@ import android.net.NetworkInfo;
 import android.os.Bundle;
 import android.util.Log;
 import android.widget.TextView;
-import com.github.pwittchen.reactivenetwork.library.rx2.ReactiveNetwork;
-import io.reactivex.android.schedulers.AndroidSchedulers;
-import io.reactivex.disposables.Disposable;
-import io.reactivex.schedulers.Schedulers;
+import com.github.pwittchen.reactivenetwork.library.rx3.ReactiveNetwork;
+import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers;
+import io.reactivex.rxjava3.disposables.Disposable;
+import io.reactivex.rxjava3.schedulers.Schedulers;
 
 public class MainActivity extends Activity {
   private static final String TAG = "ReactiveNetwork";